refactor login form to be a route
