+++ packages/lesmis/src/components/Forms/index.js
@@ -1,0 +1,6 @@
+/*
+ * Tupaia
+ * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
+ *
+ */
+export * from './LoginForm';
