+++ packages/lesmis/src/components/index.js
@@ -9,11 +9,11 @@ export * from './EntityVitalsItem';
 export * from './PartnerLogo';
 export * from './EntityVitalsItem';
 export * from './FetchLoader';
+export * from './Forms';
 export * from './Footer';
 export * from './FullPageLoader';
 export * from './Layout';
 export * from './LocationHeader';
-export * from './LoginForm';
 export * from './MainMenu';
 export * from './MapOverlaysPanel';
 export * from './MiniMap';
